+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This changelog follows the basic format outlined on [keepachangelog.com](https:/
 
 To learn more about the specifics of my changelog structure, see [doc/changelog-struct.md](doc/changelog-struct)
 
+## Unreleased - (Expected v0.6.0?)
+
+* **Added**
+  * Add a way to calculate Houndify credit usage
+
 ## v0.5.0 - Released 2021-07-06
 
 * **Added**